Refactored: libs files for better organization
