+++ src/modules/hw-app-dag/libs/transaction/tx-hash-encode.js
